Move WindowsString helper type to a separate module
